Fix sandbending sound for PK 1.8.5
- Create VersionUtil helper class for encapsulating version changes.
