+++ src/com/jedk1/jedcore/listener/AbilityListener.java
@@ -1,10 +1,10 @@
 package com.jedk1.jedcore.listener;
 
-import java.lang.reflect.Method;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
+import com.jedk1.jedcore.util.VersionUtil;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -81,30 +81,9 @@ import com.projectkorra.projectkorra.waterbending.blood.Bloodbending;
 public class AbilityListener implements Listener {
 
 	JedCore plugin;
-	private static Method isImmobilizedMethod = null;
-	
-	static {
-		try {
-			Class <?> comboClass = Class.forName("com.projectkorra.projectkorra.chiblocking.combo.ChiCombo");
-			isImmobilizedMethod = comboClass.getDeclaredMethod("isParalyzed", Player.class);
-		} catch (Exception e) {
-			
-		}
-	}
 
 	public AbilityListener(JedCore plugin) {
 		this.plugin = plugin;
-	}
-	
-	public static boolean isImmobilized(Player player) {
-		if (isImmobilizedMethod != null) {
-			try {
-				return (boolean)isImmobilizedMethod.invoke(null, player);
-			} catch (Exception e) {
-				
-			}
-		}
-		return Immobilize.isParalyzed(player);
 	}
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true) 
@@ -120,7 +99,7 @@ public class AbilityListener implements Listener {
 		if (Suffocate.isBreathbent(player)) {
 			event.setCancelled(true);
 			return;
-		} else if (Bloodbending.isBloodbent(player) || Paralyze.isParalyzed(player) || isImmobilized(player)) {
+		} else if (Bloodbending.isBloodbent(player) || Paralyze.isParalyzed(player) || VersionUtil.isImmobilized(player)) {
 			event.setCancelled(true);
 			return;
 		} else if (bPlayer.isChiBlocked()) {
@@ -291,7 +270,7 @@ public class AbilityListener implements Listener {
 			}
 		}
 
-		if (Paralyze.isParalyzed(player) || isImmobilized(player) || Bloodbending.isBloodbent(player)) {
+		if (Paralyze.isParalyzed(player) || VersionUtil.isImmobilized(player) || Bloodbending.isBloodbent(player)) {
 			event.setCancelled(true);
 			return;
 		}
